refactor: 통합 테스트 환경 수정 및 변경 반영
