+++ backend/kirikiri/src/test/java/co/kirikiri/integration/MemberIntegrationTest.java
@@ -3,8 +3,8 @@ package co.kirikiri.integration;
 import static io.restassured.RestAssured.given;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import co.kirikiri.service.dto.request.Gender;
-import co.kirikiri.service.dto.request.JoinMemberRequest;
+import co.kirikiri.service.dto.member.GenderType;
+import co.kirikiri.service.dto.member.request.JoinMemberRequest;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.time.LocalDate;
@@ -15,17 +15,16 @@ import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 
-class JoinMemberIntegrationTest extends IntegrationTest {
+class MemberIntegrationTest extends IntegrationTest {
 
     @Test
     void 정상적으로_회원가입을_성공한다() {
         //given
-        final JoinMemberRequest request = new JoinMemberRequest("ab12", "password12!@#$%", "nickname",
-            Gender.MALE, LocalDate.of(2023, Month.JULY, 12));
+        final JoinMemberRequest request = new JoinMemberRequest("ab12", "password12!@#$%", "nickname", "010-1234-5678",
+            GenderType.MALE, LocalDate.of(2023, Month.JULY, 12));
 
         //when
-        final ExtractableResponse<Response> response = joinMember(
-            request);
+        final ExtractableResponse<Response> response = 회원가입을_한다(request);
 
         //then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
@@ -36,11 +35,11 @@ class JoinMemberIntegrationTest extends IntegrationTest {
     @ValueSource(strings = {"abc", "abcdefghijklmnopqrstu"})
     void 아이디_길이가_틀린_경우_회원가입에_실패한다(final String identifier) {
         //given
-        final JoinMemberRequest request = new JoinMemberRequest(identifier, "password12!", "nickname",
-            Gender.MALE, LocalDate.of(2023, Month.JULY, 12));
+        final JoinMemberRequest request = new JoinMemberRequest(identifier, "password12!", "nickname", "010-1234-5678",
+            GenderType.MALE, LocalDate.of(2023, Month.JULY, 12));
 
         //when
-        final ExtractableResponse<Response> response = joinMember(request);
+        final ExtractableResponse<Response> response = 회원가입을_한다(request);
 
         //then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
@@ -50,11 +49,11 @@ class JoinMemberIntegrationTest extends IntegrationTest {
     @ValueSource(strings = {"Abcd", "abcdefghijklmnopqrsT", "가나다라"})
     void 아이디에_허용되지_않은_문자가_들어온_경우_회원가입에_실패한다(final String identifier) {
         //given
-        final JoinMemberRequest request = new JoinMemberRequest(identifier, "password12!", "nickname",
-            Gender.MALE, LocalDate.of(2023, Month.JULY, 12));
+        final JoinMemberRequest request = new JoinMemberRequest(identifier, "password12!", "nickname", "010-1234-5678",
+            GenderType.MALE, LocalDate.of(2023, Month.JULY, 12));
 
         //when
-        final ExtractableResponse<Response> response = joinMember(request);
+        final ExtractableResponse<Response> response = 회원가입을_한다(request);
 
         //then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
@@ -63,12 +62,12 @@ class JoinMemberIntegrationTest extends IntegrationTest {
     @Test
     void 아이디가_중복된_경우_회원가입에_실패한다() {
         //given
-        final JoinMemberRequest request = new JoinMemberRequest("ab12", "password12!", "nickname",
-            Gender.MALE, LocalDate.of(2023, Month.JULY, 12));
-        joinMember(request);
+        final JoinMemberRequest request = new JoinMemberRequest("ab12", "password12!", "nickname", "010-1234-5678",
+            GenderType.MALE, LocalDate.of(2023, Month.JULY, 12));
+        회원가입을_한다(request);
 
         //when
-        final ExtractableResponse<Response> response = joinMember(request);
+        final ExtractableResponse<Response> response = 회원가입을_한다(request);
 
         //then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
@@ -78,11 +77,11 @@ class JoinMemberIntegrationTest extends IntegrationTest {
     @ValueSource(strings = {"abcde1!", "abcdefghijklmn12"})
     void 비밀번호_길이가_틀린_경우_회원가입에_실패한다(final String password) {
         //given
-        final JoinMemberRequest request = new JoinMemberRequest("ab12", password, "nickname",
-            Gender.MALE, LocalDate.of(2023, Month.JULY, 12));
+        final JoinMemberRequest request = new JoinMemberRequest("ab12", password, "nickname", "010-1234-5678",
+            GenderType.MALE, LocalDate.of(2023, Month.JULY, 12));
 
         //when
-        final ExtractableResponse<Response> response = joinMember(request);
+        final ExtractableResponse<Response> response = 회원가입을_한다(request);
 
         //then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
@@ -92,11 +91,11 @@ class JoinMemberIntegrationTest extends IntegrationTest {
     @ValueSource(strings = {"abcdef1/", "abcdefghij1₩", "abcdefgH1!"})
     void 비밀번호에_허용되지_않은_문자가_들어온_경우_회원가입에_실패한다(final String password) {
         //given
-        final JoinMemberRequest request = new JoinMemberRequest("ab12", password, "nickname",
-            Gender.MALE, LocalDate.of(2023, Month.JULY, 12));
+        final JoinMemberRequest request = new JoinMemberRequest("ab12", password, "nickname", "010-1234-5678",
+            GenderType.MALE, LocalDate.of(2023, Month.JULY, 12));
 
         //when
-        final ExtractableResponse<Response> response = joinMember(request);
+        final ExtractableResponse<Response> response = 회원가입을_한다(request);
 
         //then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
@@ -106,11 +105,11 @@ class JoinMemberIntegrationTest extends IntegrationTest {
     @ValueSource(strings = {"abcdefgh", "abcdefghijkl"})
     void 비밀번호에_영소문자만_들어온_경우_회원가입에_실패한다(final String password) {
         //given
-        final JoinMemberRequest request = new JoinMemberRequest("ab12", password, "nickname",
-            Gender.MALE, LocalDate.of(2023, Month.JULY, 12));
+        final JoinMemberRequest request = new JoinMemberRequest("ab12", password, "nickname", "010-1234-5678",
+            GenderType.MALE, LocalDate.of(2023, Month.JULY, 12));
 
         //when
-        final ExtractableResponse<Response> response = joinMember(request);
+        final ExtractableResponse<Response> response = 회원가입을_한다(request);
 
         //then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
@@ -120,11 +119,11 @@ class JoinMemberIntegrationTest extends IntegrationTest {
     @ValueSource(strings = {"12345678", "12345678910"})
     void 비밀번호에_숫자만_들어온_경우_회원가입에_실패한다(final String password) {
         //given
-        final JoinMemberRequest request = new JoinMemberRequest("ab12", password, "nickname",
-            Gender.MALE, LocalDate.of(2023, Month.JULY, 12));
+        final JoinMemberRequest request = new JoinMemberRequest("ab12", password, "nickname", "010-1234-5678",
+            GenderType.MALE, LocalDate.of(2023, Month.JULY, 12));
 
         //when
-        final ExtractableResponse<Response> response = joinMember(request);
+        final ExtractableResponse<Response> response = 회원가입을_한다(request);
 
         //then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
@@ -134,22 +133,22 @@ class JoinMemberIntegrationTest extends IntegrationTest {
     @ValueSource(strings = {"a", "abcdefghi"})
     void 닉네임_길이가_틀린_경우_회원가입에_실패한다(final String nickname) {
         //given
-        final JoinMemberRequest request = new JoinMemberRequest("ab12", "password12!@#$%", nickname,
-            Gender.MALE, LocalDate.of(2023, Month.JULY, 12));
+        final JoinMemberRequest request = new JoinMemberRequest("ab12", "password12!@#$%", nickname, "010-1234-5678",
+            GenderType.MALE, LocalDate.of(2023, Month.JULY, 12));
 
         //when
-        final ExtractableResponse<Response> response = joinMember(request);
+        final ExtractableResponse<Response> response = 회원가입을_한다(request);
 
         //then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
     }
 
-    private ExtractableResponse<Response> joinMember(final JoinMemberRequest request) {
+    private ExtractableResponse<Response> 회원가입을_한다(final JoinMemberRequest request) {
         return given().log().all()
             .contentType(MediaType.APPLICATION_JSON_VALUE)
             .when()
             .body(request)
-            .post("/api/member/register")
+            .post("/api/member/join")
             .then()
             .log().all()
             .extract();
