+++ backend/kirikiri/src/test/resources/application.yml
@@ -20,3 +20,7 @@ logging:
           descriptor:
             sql:
               BasicBinder: TRACE
+
+server:
+  servlet:
+    contextPath: /api
